Update modern-experience-customizations.md
The section on the root site is four years out of date. All new root sites have been modern communication sites since 2019, and we now have a handy GUI function to replace an old root site with a modern site. 

I made suggested edits, with a link to the current information. Another option would be to replace this whole section with that  link, perhaps leaving the caution against deleting.